details tab created

+++ src/main/java/com/example/hotornot/model/data/local/database/models/Main.java
@@ -22,7 +22,7 @@ public class Main {
     private Double feelsLike;
     @SerializedName("temp_min")
     @Expose
-    private Integer tempMin;
+    private Double tempMin;
     @SerializedName("temp_max")
     @Expose
     private Double tempMax;
@@ -49,11 +49,11 @@ public class Main {
         this.feelsLike = feelsLike;
     }
 
-    public Integer getTempMin() {
+    public Double getTempMin() {
         return tempMin;
     }
 
-    public void setTempMin(Integer tempMin) {
+    public void setTempMin(Double tempMin) {
         this.tempMin = tempMin;
     }
